clean up, moved items to constructor, facedir mod etc.
