+++ lib/eclipse/EclipseState/Aquancon.cpp
@@ -16,73 +16,91 @@
   You should have received a copy of the GNU General Public License
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
  */
-
+#include <opm/parser/eclipse/EclipseState/Grid/FaceDir.hpp>
 #include <opm/parser/eclipse/EclipseState/Aquancon.hpp>
 #include <algorithm>
 #include <iterator>
 
 namespace Opm {
+    namespace{
+        struct AquanconRecord{
+            // Grid cell box definition to connect aquifer
+            int i1, i2, j1, j2, k1, k2;
+
+            std::vector<size_t> global_index_per_record;
+
+            // Variables constants
+            std::vector<double>  influx_coeff_per_record,  //Aquifer influx coefficient
+                                 influx_mult_per_record;   //Aquifer influx coefficient Multiplier       
+            // Cell face to connect aquifer to        
+            std::vector<int> face_per_record;           
+        };
+    }
 
     Aquancon::Aquancon(const EclipseGrid& grid, const Deck& deck)
     { 
         if (!deck.hasKeyword("AQUANCON"))
-        {
-            getAquOutput();
-        }
-        
-        else {
-            const auto& aquanconKeyword = deck.getKeyword("AQUANCON");
-            // Resize the parameter vector container based on row entries in aquancon
-            m_aqurecord.resize(aquanconKeyword.size());
-            m_aquiferID_per_record.resize(aquanconKeyword.size());
+            return;  
 
-            // We now do a loop over each record entry in aquancon
-            for (size_t aquanconRecordIdx = 0; aquanconRecordIdx < aquanconKeyword.size(); ++ aquanconRecordIdx) 
+        std::vector<Opm::AquanconRecord> m_aqurecord;
+        // Aquifer ID per record
+        std::vector<int> m_aquiferID_per_record;
+        int m_maxAquID = 0;
+
+        const auto& aquanconKeyword = deck.getKeyword("AQUANCON");
+        // Resize the parameter vector container based on row entries in aquancon
+        m_aqurecord.resize(aquanconKeyword.size());
+        m_aquiferID_per_record.resize(aquanconKeyword.size());
+
+        // We now do a loop over each record entry in aquancon
+        for (size_t aquanconRecordIdx = 0; aquanconRecordIdx < aquanconKeyword.size(); ++ aquanconRecordIdx) 
             {
-                const auto& aquanconRecord = aquanconKeyword.getRecord(aquanconRecordIdx);
-                m_aquiferID_per_record.at(aquanconRecordIdx) = aquanconRecord.getItem("AQUIFER_ID").template get<int>(0);
+            const auto& aquanconRecord = aquanconKeyword.getRecord(aquanconRecordIdx);
+            m_aquiferID_per_record.at(aquanconRecordIdx) = aquanconRecord.getItem("AQUIFER_ID").template get<int>(0);
 
-                m_aqurecord.at(aquanconRecordIdx).i1 = aquanconRecord.getItem("I1").template get<int>(0);
-                m_aqurecord.at(aquanconRecordIdx).i2 = aquanconRecord.getItem("I2").template get<int>(0);
-                m_aqurecord.at(aquanconRecordIdx).j1 = aquanconRecord.getItem("J1").template get<int>(0);
-                m_aqurecord.at(aquanconRecordIdx).j2 = aquanconRecord.getItem("J2").template get<int>(0);
-                m_aqurecord.at(aquanconRecordIdx).k1 = aquanconRecord.getItem("K1").template get<int>(0);
-                m_aqurecord.at(aquanconRecordIdx).k2 = aquanconRecord.getItem("K2").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).i1 = aquanconRecord.getItem("I1").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).i2 = aquanconRecord.getItem("I2").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).j1 = aquanconRecord.getItem("J1").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).j2 = aquanconRecord.getItem("J2").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).k1 = aquanconRecord.getItem("K1").template get<int>(0);
+            m_aqurecord.at(aquanconRecordIdx).k2 = aquanconRecord.getItem("K2").template get<int>(0);
 
-                m_aquiferID_per_record.at(aquanconRecordIdx) = aquanconRecord.getItem("AQUIFER_ID").template get<int>(0);
-                m_maxAquID = (m_maxAquID < m_aquiferID_per_record.at(aquanconRecordIdx) )?
-                                m_aquiferID_per_record.at(aquanconRecordIdx) : m_maxAquID;
+            m_aquiferID_per_record.at(aquanconRecordIdx) = aquanconRecord.getItem("AQUIFER_ID").template get<int>(0);
+            m_maxAquID = (m_maxAquID < m_aquiferID_per_record.at(aquanconRecordIdx) )?
+                            m_aquiferID_per_record.at(aquanconRecordIdx) : m_maxAquID;
               
-                double influx_coeff = aquanconRecord.getItem("INFLUX_COEFF").getSIDouble(0);
+            double influx_coeff = aquanconRecord.getItem("INFLUX_COEFF").getSIDouble(0);
 
-                double influx_mult = aquanconRecord.getItem("INFLUX_MULT").getSIDouble(0);
+            double influx_mult = aquanconRecord.getItem("INFLUX_MULT").getSIDouble(0);
 
-                std::string faceString = aquanconRecord.getItem("FACE").getTrimmedString(0);
+            FaceDir::DirEnum faceDir = FaceDir::FromString(aquanconRecord.getItem("FACE").getTrimmedString(0));
               
-                // Loop over the cartesian indices to convert to the global grid index
-                for (int k=m_aqurecord.at(aquanconRecordIdx).k1; k <= m_aqurecord.at(aquanconRecordIdx).k2; k++) {
-                    for (int j=m_aqurecord.at(aquanconRecordIdx).j1; j <= m_aqurecord.at(aquanconRecordIdx).j2; j++)
-                        for (int i=m_aqurecord.at(aquanconRecordIdx).i1; i <= m_aqurecord.at(aquanconRecordIdx).i2; i++)
-                            m_aqurecord.at(aquanconRecordIdx).global_index_per_record.push_back
-                                                                (grid.getGlobalIndex(i-1 ,j-1 ,k- 1)
-                                                                );
-                }
-                size_t global_index_per_record_size = m_aqurecord.at(aquanconRecordIdx).global_index_per_record.size();
-                m_aqurecord.at(aquanconRecordIdx).influx_coeff_per_record.resize(global_index_per_record_size,influx_coeff);
-                m_aqurecord.at(aquanconRecordIdx).influx_mult_per_record.resize(global_index_per_record_size,influx_mult);
-                m_aqurecord.at(aquanconRecordIdx).face_per_record.resize(global_index_per_record_size,faceString);
+            // Loop over the cartesian indices to convert to the global grid index
+            for (int k=m_aqurecord.at(aquanconRecordIdx).k1; k <= m_aqurecord.at(aquanconRecordIdx).k2; k++) {
+                for (int j=m_aqurecord.at(aquanconRecordIdx).j1; j <= m_aqurecord.at(aquanconRecordIdx).j2; j++)
+                    for (int i=m_aqurecord.at(aquanconRecordIdx).i1; i <= m_aqurecord.at(aquanconRecordIdx).i2; i++)
+                        m_aqurecord.at(aquanconRecordIdx).global_index_per_record.push_back
+                                                            (grid.getGlobalIndex(i-1 ,j-1 ,k- 1)
+                                                            );
             }
-
-            // Collate_function
-            collate_function(m_aquoutput);
-
-            // Logic for grid connection applied here
-            logic_application(m_aquoutput);
+            size_t global_index_per_record_size = m_aqurecord.at(aquanconRecordIdx).global_index_per_record.size();
+            m_aqurecord.at(aquanconRecordIdx).influx_coeff_per_record.resize(global_index_per_record_size,influx_coeff);
+            m_aqurecord.at(aquanconRecordIdx).influx_mult_per_record.resize(global_index_per_record_size,influx_mult);
+            m_aqurecord.at(aquanconRecordIdx).face_per_record.resize(global_index_per_record_size,faceDir);
         }
-    }                                          
+
+        // Collate_function
+        collate_function(m_aquoutput, m_aqurecord, m_aquiferID_per_record, m_maxAquID);
+
+        // Logic for grid connection applied here
+        logic_application(m_aquoutput);
+    }
+                                         
     
     // This function is used to convert from a per record vector to a per aquifer ID vector.
-    void Aquancon::collate_function(std::vector<Aquancon::AquanconOutput>& output_vector)
+    void Aquancon::collate_function(std::vector<Aquancon::AquanconOutput>& output_vector, 
+                                    std::vector<Opm::AquanconRecord>& m_aqurecord, 
+                                    std::vector<int> m_aquiferID_per_record, int m_maxAquID)
     {
         output_vector.resize(m_maxAquID);
         // Find record indices at which the aquifer ids are located in
@@ -90,7 +108,7 @@ namespace Opm {
         {
             std::vector<int> result_id;
 
-            convert_record_id_to_aquifer_id(result_id, i);
+            convert_record_id_to_aquifer_id(result_id, i, m_aquiferID_per_record);
 
             // We add the aquifer id into each element of output_vector
             output_vector.at(i - 1).aquiferID = i;
@@ -138,7 +156,8 @@ namespace Opm {
         //TODO: Total number of grid blocks connected to aquifer must not exceed item 6 of AQUDIMS
     }
 
-    void Aquancon::convert_record_id_to_aquifer_id(std::vector<int>& record_indices_matching_id, int i)
+    void Aquancon::convert_record_id_to_aquifer_id(std::vector<int>& record_indices_matching_id,
+                                                   int i, std::vector<int> m_aquiferID_per_record)
     {
         auto it = std::find_if( m_aquiferID_per_record.begin(), m_aquiferID_per_record.end(), 
                                     [&](int id) {
