+++ lib/eclipse/include/opm/parser/eclipse/EclipseState/Aquancon.hpp
@@ -32,11 +32,11 @@
 #include <opm/parser/eclipse/Deck/DeckItem.hpp>
 #include <opm/parser/eclipse/Deck/DeckRecord.hpp>
 #include <opm/parser/eclipse/Deck/DeckKeyword.hpp>
-#include <boost/concept_check.hpp>
-
-#include <iostream>
 
 namespace Opm {
+    namespace{
+        struct AquanconRecord;
+        }
 
     class Aquancon {
         public:
@@ -46,7 +46,7 @@ namespace Opm {
                 std::vector<size_t> global_index;
                 std::vector<double> influx_coeff; // Size = size(global_index)
                 std::vector<double> influx_multiplier; // Size = size(global_index)
-                std::vector<std::string> reservoir_face_dir; // Size = size(global_index)
+                std::vector<int> reservoir_face_dir; // Size = size(global_index)
             };
 
             Aquancon(const EclipseGrid& grid, const Deck& deck);
@@ -55,33 +55,16 @@ namespace Opm {
     
         private:
 
-            struct AquanconRecord{
-                    // Grid cell box definition to connect aquifer
-                    int i1, i2, j1, j2, k1, k2;
-
-                    std::vector<size_t> global_index_per_record;
-
-                    // Variables constants
-                    std::vector<double>  influx_coeff_per_record,  //Aquifer influx coefficient
-                                         influx_mult_per_record;   //Aquifer influx coefficient Multiplier       
-                    // Cell face to connect aquifer to        
-                    std::vector<std::string> face_per_record;           
-
-            };
-
             void logic_application(std::vector<Aquancon::AquanconOutput>& output_vector);
 
-            void collate_function(std::vector<Aquancon::AquanconOutput>& output_vector);
+            void collate_function(std::vector<Aquancon::AquanconOutput>& output_vector, 
+                                  std::vector<Opm::AquanconRecord>& m_aqurecord, 
+                                  std::vector<int> m_aquiferID_per_record, int m_maxAquID);
 
-            void convert_record_id_to_aquifer_id(std::vector<int>& record_indices_matching_id, int i);
+            void convert_record_id_to_aquifer_id(std::vector<int>& record_indices_matching_id, int i,
+                                                 std::vector<int> m_aquiferID_per_record);
 
             std::vector<Aquancon::AquanconOutput> m_aquoutput;
-
-            std::vector<Aquancon::AquanconRecord> m_aqurecord;
-
-            // Aquifer ID
-            std::vector<int> m_aquiferID_per_record;
-            int m_maxAquID = 0;
     };
 }
